requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
groq
chromadb
sentence-transformers
openai
httpx
pytest
pytest-asyncio
python-multipart
pydantic[email]